Refactors models and creates tests

+++ backend/src/main/java/com/gabrielmbarboza/sienge/model/Transport.java
@@ -1,5 +1,5 @@
 package com.gabrielmbarboza.sienge.model;
 
 public interface Transport {
-  public Double calculateTotalCost(Vehicle vehicle, int kmTraveled, int weightCargo);
+  public Double calculateTotalCost(Vehicle vehicle, Double kmTraveled, Double weightCargo);
 }
